requirements.txt
streamlit>=1.26.0
pdfplumber>=0.9.0
llama-index>=0.10.45
llama-index-llms-gemini>=0.1.0
google-generativeai>=0.6.0
python-dotenv>=1.0.1
tqdm>=4.66.5
transformers>=4.45.0
torch>=2.3.0
spacy>=3.6.0
